drivers/sensor: lis2mdl: Fix temperature sample handling
The lis2mdl temperature samples work with a level of 25 Celsius.
When temperature goes below that level the samples become negative
and there was an issue in properly propagating the sign.

Fix #35910

+++ drivers/sensor/lis2mdl/lis2mdl.h
@@ -56,7 +56,7 @@ struct lis2mdl_data {
 	const struct device *bus;
 	uint16_t i2c_addr;
 	int16_t mag[3];
-	int32_t temp_sample;
+	int16_t temp_sample;
 
 	stmdev_ctx_t *ctx;
